# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kimdasa Construction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /market-prices:
    get:
      operationId: getMarketPrices
      tags: [market-prices]
      summary: List market prices (public)
      responses:
        "200":
          description: Market prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketPrice"
    post:
      operationId: createMarketPrice
      tags: [market-prices]
      summary: Create a market price entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketPriceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPrice"
  /market-prices/staleness:
    get:
      operationId: getMarketPriceStaleness
      tags: [market-prices]
      summary: Get stale price entries (older than 30 days)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Staleness report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StalenessReport"
  /market-prices/{id}:
    patch:
      operationId: updateMarketPrice
      tags: [market-prices]
      summary: Update a market price
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMarketPriceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPrice"
    delete:
      operationId: deleteMarketPrice
      tags: [market-prices]
      summary: Delete a market price
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /virtual-agents:
    get:
      operationId: getVirtualAgents
      tags: [virtual-agents]
      summary: List virtual agents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualAgent"
    post:
      operationId: createVirtualAgent
      tags: [virtual-agents]
      summary: Create a virtual agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVirtualAgentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAgent"
  /virtual-agents/{id}:
    patch:
      operationId: updateVirtualAgent
      tags: [virtual-agents]
      summary: Update a virtual agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVirtualAgentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualAgent"
...
  /site-config/{key}:
    put:
      operationId: updateSiteConfig
      tags: [site-config]
      summary: Update a site config key
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteConfigRequest"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Task:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        jobId:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - createdAt
        - updatedAt
    CreateTaskRequest:
      type: object
      properties:
        leadId:
          type: integer
        customerId:
          type: integer
        jobId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        assignedTo:
          type: string
      required:
        - title
    MarketPrice:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        region:
          type: string
          enum: [new_jersey, pennsylvania]
        priceType:
          type: string
          enum: [installed, material, labor]
        unit:
          type: string
        minPrice:
          type: string
          nullable: true
        avgPrice:
          type: string
          nullable: true
        premiumPrice:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - priceType
        - region
        - unit
        - updatedAt
        - createdAt
    CreateMarketPriceRequest:
      type: object
      properties:
        category:
          type: string
        region:
          type: string
          enum: [new_jersey, pennsylvania]
        priceType:
          type: string
          enum: [installed, material, labor]
        unit:
          type: string
        minPrice:
          type: string
        avgPrice:
          type: string
        premiumPrice:
          type: string
        notes:
          type: string
        source:
          type: string
      required:
        - category
        - region
    StalenessReport:
      type: object
      properties:
        staleCount:
          type: integer
        staleItems:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              category:
                type: string
              region:
                type: string
              updatedAt:
                type: string
                format: date-time
      required:
        - staleCount
        - staleItems
...
    VirtualAgent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        agentType:
          type: string
        systemPrompt:
          type: string
          nullable: true
        language:
          type: string
        isActive:
          type: boolean
        config:
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - agentType
        - language
        - isActive
        - createdAt
        - updatedAt
    CreateVirtualAgentRequest:
      type: object
      properties:
        name:
          type: string
        agentType:
          type: string
        systemPrompt:
          type: string
        language:
          type: string
        isActive:
          type: boolean
        config:
          type: object
      required:
        - name
        - agentType
    SiteConfig:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - key
        - updatedAt
    UpdateSiteConfigRequest:
      type: object
      properties:
        value:
          type: string
        description:
          type: string
...